+++ homematicip/base/enums.py
@@ -202,6 +202,7 @@ class DeviceType(AutoNameEnum):
     ACCELERATION_SENSOR = auto()
     ALARM_SIREN_INDOOR = auto()
     ALARM_SIREN_OUTDOOR = auto()
+    BLIND_MODULE = auto()
     BRAND_BLIND = auto()
     BRAND_DIMMER = auto()
     BRAND_PUSH_BUTTON = auto()
@@ -221,6 +222,7 @@ class DeviceType(AutoNameEnum):
     HEATING_SWITCH_2 = auto()
     HEATING_THERMOSTAT = auto()
     HEATING_THERMOSTAT_COMPACT = auto()
+    HOME_CONTROL_ACCESS_POINT = auto()
     HOERMANN_DRIVES_MODULE = auto()
     KEY_REMOTE_CONTROL_4 = auto()
     KEY_REMOTE_CONTROL_ALARM = auto()
@@ -383,6 +385,7 @@ class WindValueType(AutoNameEnum):
 class FunctionalChannelType(AutoNameEnum):
     FUNCTIONAL_CHANNEL = auto()
     ACCELERATION_SENSOR_CHANNEL = auto()
+    ACCESS_CONTROLLER_CHANNEL = auto()
     ALARM_SIREN_CHANNEL = auto()
     ANALOG_OUTPUT_CHANNEL = auto()
     ANALOG_ROOM_CONTROL_CHANNEL = auto()
@@ -417,6 +420,7 @@ class FunctionalChannelType(AutoNameEnum):
     PASSAGE_DETECTOR_CHANNEL = auto()
     PRESENCE_DETECTION_CHANNEL = auto()
     ROTARY_HANDLE_CHANNEL = auto()
+    SHADING_CHANNEL = auto()
     SHUTTER_CHANNEL = auto()
     SHUTTER_CONTACT_CHANNEL = auto()
     SINGLE_KEY_CHANNEL = auto()
@@ -544,3 +548,30 @@ class ConnectionType(AutoNameEnum):
     HMIP_WIRED = auto()
     HMIP_LAN = auto()
     HMIP_WLAN = auto()
+
+
+class ShadingStateType(AutoNameEnum):
+    NOT_POSSIBLE = auto()
+    NOT_EXISTENT = auto()
+    POSITION_USED = auto()
+    TILT_USED = auto()
+    NOT_USED = auto()
+    MIXED = auto()
+
+
+class DriveSpeed(AutoNameEnum):
+    CREEP_SPEED = auto()
+    SLOW_SPEED = auto()
+    NOMINAL_SPEED = auto()
+    OPTIONAL_SPEED = auto()
+
+
+class ShadingPackagePosition(AutoNameEnum):
+    LEFT = auto()
+    RIGHT = auto()
+    CENTER = auto()
+    SPLIT = auto()
+    TOP = auto()
+    BOTTOM = auto()
+    TDBU = auto()
+    NOT_USED = auto()