+++ homematicip/device.py
@@ -198,6 +198,23 @@ class Device(HomeMaticIPObject):
         return fc
 
 
+class HomeControlAccessPoint(Device):
+    def __init__(self, connection):
+        super().__init__(connection)
+        self.dutyCycleLevel = 0.0
+        self.accessPointPriority = 0
+        self.signalBrightness = 0
+        self._baseChannel = "ACCESS_CONTROLLER_CHANNEL"
+
+    def from_json(self, js):
+        super().from_json(js)
+        c = get_functional_channel(self._baseChannel, js)
+        if c:
+            self.set_attr_from_dict("dutyCycleLevel", c)
+            self.set_attr_from_dict("accessPointPriority", c)
+            self.set_attr_from_dict("signalBrightness", c)
+
+
 class SabotageDevice(Device):
     def __init__(self, connection):
         super().__init__(connection)
@@ -1188,6 +1205,91 @@ class FullFlushBlind(FullFlushShutter):
 
 class BrandBlind(FullFlushBlind):
     """ HMIP-BBL (Blind Actuator for brand switches) """
+
+
+class BlindModule(Device):
+    """ HMIP-HDM1 (Hunter Douglas & erfal window blinds) """
+
+    def __init__(self, connection):
+        super().__init__(connection)
+        self.automationDriveSpeed = DriveSpeed.CREEP_SPEED
+        self.manualDriveSpeed = DriveSpeed.CREEP_SPEED
+        self.favoritePrimaryShadingPosition = 0.0
+        self.favoriteSecondaryShadingPosition = 0.0
+        self.primaryShadingLevel = 0.0
+        self.secondaryShadingLevel = 0.0
+        self.previousPrimaryShadingLevel = 0.0
+        self.previousSecondaryShadingLevel = 0.0
+        self.identifyOemSupported = False
+        self.productId = 0
+        self.primaryCloseAdjustable = False
+        self.primaryOpenAdjustable = False
+        self.primaryShadingStateType = ShadingStateType.NOT_EXISTENT
+        self.primaryCloseAdjustable = False
+        self.primaryOpenAdjustable = False
+        self.primaryShadingStateType = ShadingStateType.NOT_EXISTENT
+        self.profileMode = ProfileMode.MANUAL
+        self.userDesiredProfileMode = ProfileMode.MANUAL
+        self.processing = False
+        self.shadingDriveVersion = None
+        self.shadingPackagePosition = ShadingPackagePosition.NOT_USED
+        self.shadingPositionAdjustmentActive = None
+        self.shadingPositionAdjustmentClientId = None
+
+    def from_json(self, js):
+        super().from_json(js)
+        c = get_functional_channel("SHADING_CHANNEL", js)
+        if c:
+            self.set_attr_from_dict("automationDriveSpeed", c, DriveSpeed)
+            self.set_attr_from_dict("manualDriveSpeed", c, DriveSpeed)
+
+            self.set_attr_from_dict("favoritePrimaryShadingPosition", c)
+            self.set_attr_from_dict("favoriteSecondaryShadingPosition", c)
+
+            self.set_attr_from_dict("primaryCloseAdjustable", c)
+            self.set_attr_from_dict("primaryOpenAdjustable", c)
+            self.set_attr_from_dict("primaryShadingStateType", c, ShadingStateType)
+            self.set_attr_from_dict("secondaryCloseAdjustable", c)
+            self.set_attr_from_dict("secondaryOpenAdjustable", c)
+            self.set_attr_from_dict("secondaryShadingStateType", c, ShadingStateType)
+
+            self.set_attr_from_dict("primaryShadingLevel", c)
+            self.set_attr_from_dict("secondaryShadingLevel", c)
+
+            self.set_attr_from_dict("previousPrimaryShadingLevel", c)
+            self.set_attr_from_dict("previousSecondaryShadingLevel", c)
+
+            self.set_attr_from_dict("identifyOemSupported", c)
+            self.set_attr_from_dict("productId", c)
+
+            self.set_attr_from_dict("profileMode", c, ProfileMode)
+            self.set_attr_from_dict("userDesiredProfileMode", c, ProfileMode)
+
+            self.set_attr_from_dict("shadingDriveVersion", c)
+            self.set_attr_from_dict("shadingPackagePosition", c, ShadingPackagePosition)
+            self.set_attr_from_dict("shadingPositionAdjustmentActive", c)
+            self.set_attr_from_dict("shadingPositionAdjustmentClientId", c)
+
+    def set_primary_shading_level(self, primaryShadingLevel: float):
+        data = {
+            "channelIndex": 1,
+            "deviceId": self.id,
+            "primaryShadingLevel": primaryShadingLevel,
+        }
+        return self._restCall("device/control/setPrimaryShadingLevel", json.dumps(data))
+
+    def set_secondary_shading_level(
+        self, primaryShadingLevel: float, secondaryShadingLevel: float
+    ):
+        data = {
+            "channelIndex": 1,
+            "deviceId": self.id,
+            "primaryShadingLevel": primaryShadingLevel,
+            "secondaryShadingLevel": secondaryShadingLevel,
+        }
+        return self._restCall(
+            "device/control/setSecondaryShadingLevel", json.dumps(data)
+        )
 
 
 class LightSensor(Device):