+++ homematicip_demo/fake_cloud_server.py
@@ -16,7 +16,7 @@ class AsyncFakeCloudServer:
     def __init__(self, home_path=Path(__file__).parent.joinpath("json_data/home.json")):
         self.reset(home_path=home_path)
 
-    def reset(self,home_path=Path(__file__).parent.joinpath("json_data/home.json")):
+    def reset(self, home_path=Path(__file__).parent.joinpath("json_data/home.json")):
         with open(home_path, encoding="utf-8") as file:
             self.data = json.load(file, encoding="UTF-8")
             self.sgtin = "3014F711A000000BAD0C0DED"
@@ -34,8 +34,8 @@ class AsyncFakeCloudServer:
 
             self.client_auth_waiting = None  # used in auth
             self.home_id = "00000000-0000-0000-0000-000000000001"
-            self.ws = None   
-     
+            self.ws = None
+
     def __del__(self):
         if self.ws:
             self.ws.close()  # pragma: no cover
@@ -536,6 +536,39 @@ class AsyncFakeCloudServer:
         d = self.data["devices"][js["deviceId"]]
         channelIndex = str(js["channelIndex"])
         d["functionalChannels"][channelIndex]["shutterLevel"] = js["shutterLevel"]
+
+        return response
+
+    @validate_authorization
+    async def post_hmip_device_control_setPrimaryShadingLevel(
+        self, request: web.Request
+    ) -> web.Response:
+
+        response = web.json_response(None)
+        js = json.loads(request.data)
+        d = self.data["devices"][js["deviceId"]]
+        channelIndex = str(js["channelIndex"])
+        d["functionalChannels"][channelIndex]["primaryShadingLevel"] = js[
+            "primaryShadingLevel"
+        ]
+
+        return response
+
+    @validate_authorization
+    async def post_hmip_device_control_setSecondaryShadingLevel(
+        self, request: web.Request
+    ) -> web.Response:
+
+        response = web.json_response(None)
+        js = json.loads(request.data)
+        d = self.data["devices"][js["deviceId"]]
+        channelIndex = str(js["channelIndex"])
+        d["functionalChannels"][channelIndex]["primaryShadingLevel"] = js[
+            "primaryShadingLevel"
+        ]
+        d["functionalChannels"][channelIndex]["secondaryShadingLevel"] = js[
+            "secondaryShadingLevel"
+        ]
 
         return response
 
@@ -1063,7 +1096,7 @@ class AsyncFakeCloudServer:
         with open(Path(__file__).parent.joinpath("json_data", js["file"])) as file:
             self.data = json.load(file, encoding="UTF-8")
         return web.json_response(None)
-    
+
     # endregion
 
     # region websocket