+++ tests/test_devices.py
@@ -1373,3 +1373,14 @@ def test_wall_thermostat_basic(fake_home: Home):
             "actualTemperature(16.0) humidity(42) vaporAmount(5.710127947243264) "
             "setPointTemperature(12.0)"
         )
+
+
+def test_home_control_access_point(fake_home: Home):
+    with no_ssl_verification():
+        d = fake_home.search_device_by_id("3014F711A000000BAD0C0DED")
+
+        assert str(d) == (
+            "HmIP-HAP HOME_CONTROL_ACCESS_POINT lowBat(None) unreach(False) rssiDeviceValue(None) "
+            "rssiPeerValue(None) configPending(False) dutyCycle(False) dutyCycleLevel(8.0) "
+            "accessPointPriority(1) signalBrightness(1.0)"
+        )